Switch to new 1.2GHz kernel with WiFi fix (1/2)
This one is from ChevyNo1.  This one was reported
to have working WiFi.  I believe this is also
a Froyo-based kernel.

This commit consolidates all kernel related stuff
into one place, rather than in both device AND
vendor.

This needs tested, and if its confirmed working in
all aspects, this will be merged.

Change-Id: I63df2f1c539f4a231d7c7974ddc56d8742d5db6a